Split out `tokio-trace-core`
