+++ tokio-trace-core/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "tokio-trace-core"
+version = "0.1.0"
+authors = ["Eliza Weisman <eliza@buoyant.io>"]
+
+[features]
+default = []
+test-support = []
